server/requirements.txt
anyio==3.6.2
asgiref==3.5.2
Babel==2.11.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cohere==3.1.3
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
greenlet==2.0.1
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.2
httptools==0.5.0
httpx==0.23.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
orjson==3.8.2
psycopg==3.1.4
psycopg-binary==3.1.4
psycopg2-binary==2.9.5
py-moneyed==2.0
pydantic==1.10.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.6
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.45
sqlalchemy-cockroachdb==1.4.4
sqlparse==0.4.3
starlette==0.22.0
typing_extensions==4.4.0
ujson==5.6.0
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
Werkzeug==2.2.2
zipp==3.11.0
